Extract InlineTabBar
